+++ lib/generators/rodauth/templates/app/models/account.rb
@@ -1,2 +1,3 @@
 class Account < ApplicationRecord
+  include Rodauth::Rails.model
 end