+++ test/rails_app/app/models/account.rb
@@ -1,2 +1,3 @@
 class Account < ApplicationRecord
+  validates_presence_of :email
 end